requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14
django-environ>=0.11
dj-database-url>=2.1
psycopg[binary]>=3.1
stripe>=10.0
celery>=5.3
redis>=5.0
python-jose[cryptography]>=3.3
requests>=2.31
boto3>=1.34
sentry-sdk>=1.39
django-htmx>=1.17
agora-token-builder>=1.0.0
whitenoise>=6.6
